backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
openai
python-dotenv
supabase
